feat: added stack for nested statements
WORK IN PROGRESS

The parse tree will still be incorrect as operations after the nested statement may not be resolved properly. Take a and (b or c) and x

and *tmp_node
|-> a
|-> or
|->|->b
|->|->c

and x is then added to tmp_node, this will fail. therefore the first operation past the query should be resolved before the query such that the root of this example tree would have been and