feat(KDD): added re2 regex engine
The re2 regex engine will be replacing regex.h as it has better properties - namely linear time complexity guarantees. Whilst it is worse in some conditions, it does not have the exponential worst case scenario. In the interest of not being DoSed I think it is a wise change. Also it is cross platform so is another step forwards in getting this running on WinDos